+++ App/Client/FrmActive.h
@@ -28,15 +28,19 @@ public Q_SLOTS:
     void slotViewChanged(const QWidget* pView);
 Q_SIGNALS:
     void sigChanged(COperate* c);
+    void sigAddToFavorite(const QString& szName, const QString& szDescription, const QIcon& icon, const QString &szFile);
+
 private Q_SLOTS:
     void slotCustomContextMenu(const QPoint &pos);
     void slotClicked(const QModelIndex& index);
+    void slotAddToFavorite();
 private:
     QVector<COperate*> &m_Operates;
     CParameterApp &m_ParameterApp;
     QMenu* m_pOperate;
     QAction* m_pStop;
     RabbitCommon::CRecentMenu* m_pRecentMenu;
+    QAction* m_pAddToFavorite;
     QTableView* m_pTableView;
     QStandardItemModel* m_pModel;
     QMenu* m_pMenu;